swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field92K.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(rateTypeCode)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field92K
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field92K(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getRateTypeCode() {
        return getComponent2();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", rateTypeCode=" + getComponent2() ).append( "]").toString();
    }

}
